requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0
tqdm>=4.65.0
pyyaml>=6.0.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
opencv-python>=4.8.0
albumentations>=1.3.0  # for image augmentations
wandb>=0.15.0  # for experiment tracking
scipy>=1.7.0  # for scientific computing
timm>=0.9.0  # for vision models and utilities
pylatex>=1.4.1  # for LaTeX document generation
latexcodec>=2.0.1  # for LaTeX encoding/decoding